+++ src/Service/Rekognition/src/ValueObject/HumanLoopConfig.php
@@ -1,0 +1,84 @@
+<?php
+
+namespace AsyncAws\Rekognition\ValueObject;
+
+use AsyncAws\Core\Exception\InvalidArgument;
+
+/**
+ * Sets up the configuration for human evaluation, including the FlowDefinition the image will be sent to.
+ */
+final class HumanLoopConfig
+{
+    /**
+     * The name of the human review used for this image. This should be kept unique within a region.
+     */
+    private $humanLoopName;
+
+    /**
+     * The Amazon Resource Name (ARN) of the flow definition. You can create a flow definition by using the Amazon Sagemaker
+     * CreateFlowDefinition Operation.
+     *
+     * @see https://docs.aws.amazon.com/sagemaker/latest/dg/API_CreateFlowDefinition.html
+     */
+    private $flowDefinitionArn;
+
+    /**
+     * Sets attributes of the input data.
+     */
+    private $dataAttributes;
+
+    /**
+     * @param array{
+     *   HumanLoopName: string,
+     *   FlowDefinitionArn: string,
+     *   DataAttributes?: null|HumanLoopDataAttributes|array,
+     * } $input
+     */
+    public function __construct(array $input)
+    {
+        $this->humanLoopName = $input['HumanLoopName'] ?? null;
+        $this->flowDefinitionArn = $input['FlowDefinitionArn'] ?? null;
+        $this->dataAttributes = isset($input['DataAttributes']) ? HumanLoopDataAttributes::create($input['DataAttributes']) : null;
+    }
+
+    public static function create($input): self
+    {
+        return $input instanceof self ? $input : new self($input);
+    }
+
+    public function getDataAttributes(): ?HumanLoopDataAttributes
+    {
+        return $this->dataAttributes;
+    }
+
+    public function getFlowDefinitionArn(): string
+    {
+        return $this->flowDefinitionArn;
+    }
+
+    public function getHumanLoopName(): string
+    {
+        return $this->humanLoopName;
+    }
+
+    /**
+     * @internal
+     */
+    public function requestBody(): array
+    {
+        $payload = [];
+        if (null === $v = $this->humanLoopName) {
+            throw new InvalidArgument(sprintf('Missing parameter "HumanLoopName" for "%s". The value cannot be null.', __CLASS__));
+        }
+        $payload['HumanLoopName'] = $v;
+        if (null === $v = $this->flowDefinitionArn) {
+            throw new InvalidArgument(sprintf('Missing parameter "FlowDefinitionArn" for "%s". The value cannot be null.', __CLASS__));
+        }
+        $payload['FlowDefinitionArn'] = $v;
+        if (null !== $v = $this->dataAttributes) {
+            $payload['DataAttributes'] = $v->requestBody();
+        }
+
+        return $payload;
+    }
+}